Add reusable auto-translate documentation workflow (#29)
* Create auto-translate-documentation-reusable.yml

* Create sample-usage.yaml

* Create README.md

* Change folder name

* Change workflow name

* Remove `pull_request` and `workflow_dispatch` blocks

The `pull_request` and `workflow_dispatch` blocks aren't necessary since we're calling this workflow from other repositories. In those other repositories, these blocks are necessary and already included, as shown in the workflow example (`auto-translate-docs-script/sample-usage.yaml`).

Co-Authored-By: Vincent Guilpain <3835021+Torch3333@users.noreply.github.com>

* Make workflow run only on merge

The workflow shouldn't run when closed. No PR would be created because the action in this ("actions") repository specifies to not run on closed PRs, but not specifying that in this file would waste one minute of GitHub action time.

* Specify `workflow_dispatch` in `if` statement

Added as recommended in https://github.com/scalar-labs/docs-internal-scalardb/pull/1477#discussion_r2314786900

Co-Authored-By: Vincent Guilpain <3835021+Torch3333@users.noreply.github.com>

* Update comment

---------

Co-authored-by: Vincent Guilpain <3835021+Torch3333@users.noreply.github.com>